movies_admin/requirements/dev.txt
-r base.txt
gunicorn==20.0.4 # Боевой http-сервер для python-приложений
python-dotenv==0.15.0
django-debug-toolbar==3.1.1 # выдвигающееся меню с результатами анализа загрузки страницы
django-extensions==3.0.9
django-cors-headers==3.5.0